+++ AssemblyGenerator/Core.cs
@@ -38,15 +38,14 @@ namespace MelonLoader.AssemblyGenerator
                     var hash = md5.ComputeHash(stream);
                     CurrentGameAssemblyHash = BitConverter.ToString(hash).Replace("-", "").ToLowerInvariant();
             }
+            SamboyAPI.Setup();
             unitydependencies = new UnityDependencies();
             dumper = new Il2CppDumper();
             il2cppassemblyunhollower = new Il2CppAssemblyUnhollower();
             deobfuscationMap = new DeobfuscationMap();
         }
 
-        private static int Run() => Run(null);
-
-        private static int Run(string nullarg)
+        private static int Run()
         {
             if (!unitydependencies.Download()
                 || !dumper.Download()